+++ docs/content/patterns/conditional-reveal/index.mdx
@@ -3,6 +3,8 @@ title: Conditionally revealed questions
 group: Patterns
 description: Conditionally reveal a related question when a user selects a specific radio or checkbox option. This ensures that users only encounter questions when it is applicable to their selection.
 version: 1.0.0
+storybookPath: /story/patterns-conditionally-revealed-questions--radios
+githubTemplatePath: .storybook/stories/ConditionalReveal.tsx
 ---
 
 <DoHeading />
