+++ packages/react/src/control-input/docs/overview.mdx
@@ -2,7 +2,7 @@
 title: Control input
 description: Control inputs help users select one or more options from a list. Our control inputs consist of checkboxes and radio buttons.
 group: Forms
-storybookPath: /story/forms-checkbox--basic
+storybookPath: /story/forms-controlgroup--checkbox-group
 figmaGalleryNodeId: 12926%3A104981
 relatedComponents: ['switch']
 ---
